requirements.txt
langchain
langchain-openai
python-dotenv==1.0.1
langchain-groq
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
faiss-cpu
sentence-transformers
langchain_huggingface 
chromadb
langchain_chroma
streamlit
langchain-ollama
fastapi
uvicorn
langserve
sse_starlette